pyproject.toml
[project]
name = "keepassxc-run"
version = "0.0.2"
description = "KeePassXC runner"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.scripts]
keepassxc-run = "keepassxc_run.cli:main"

[dependency-groups]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.11.1",
    "ruff>=0.6.8",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "require_db: tests which require KeePassXC database)",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    # exclude main function
    '''if __name__ == ['"]__main__['"]'''
]
